backend/requirements.txt
asgiref==3.3.4
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-probes==1.6.0
djangorestframework==3.12.4
Faker==8.5.1
Markdown==3.3.4
psycopg2==2.8.6
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
